Use the correct IOC
That line of code was unfortunately wrong. The IDataStore is registered
in the App.Services container (the one that we created, not Xamarin).
The DependencyServices are the IOC created and used by Xamarin/Android.
We're not using it because it's hard to use in tests and it's otherwise
quite limited.